Code changes to fix the issue PP-228
PP-228 Description: PBS Server is restarting Python interpreter too often

Solution: As we updated PBS to use Python 2.7.3 which is the latest and
last release of Python 2.x and it also fixes several memory leaks which
were there in Python 2.5.1, we NO longer need to restart Python Interpreter
for every 100 hook events to service or PBS Python objects reached to 1000.
So the resolution for this issue is to remove the code which does these
conditional checks.