Import hook sets 'output-file' attribute
When a hook script is imported, an 'output-file' attribute is added to
the operation's attribute list.  The purpose of this additional
attribute is to provide any manangement hook scripts with read access
to the actual hook script that was just imported and verified.  The
'output-file' attribute is only added if the script is successfully
imported.