Only attempt to extend job script directive lines (#2356)
* Only attempt to extend job script directive lines

PBS job scripts should only extend PBS directives in the script. Line
extensions in the rest of the script can change the meaning of the job
script. For example, a comment line in a shell script that ended with
'\' would cause the next line of the script to be commented out if the
line is processed with the line extension code.

* Trust libc to be standards compliant and free(NULL) is noop