New end-job hook tests and common code refactoring
New end-job hooks tests have been added.  Four of these tests currently
fail, three of which appear to be an issue with the end-job hook
additions not running end-job hooks during a forced delete.  The fourth
failure may be a bug in the qdel or server code.  More investigation is
needed.

Supporting methods have also been added to the TestHookEndJob class to
provide support for the additional tests.