DBPRT to log file
At times, it is useful to have the debugging printf (DBPRT) messages
sent to the log files rather than stdout so that they are properly
interspersed with the other log messages.  When DBPRT_LOG is defined, an
alternative implementation of the DBPRT macro is used.  This
implementation uses the logging module to log the message at the
PBSEVENT_DEBUG3 level.  This commit also contains a few fixes to code
using DBPRT.