Eliminate cyclic references in the pbs module
The server classes, as well as the PbsAttributeDescriptor and
pbs_resource classes, each contained dictionaries that tracked objects
instantiated from the class.  This resulted in a cyclic reference that
that prevented the garbage collector from deleting the objects and
memory use to grow every time a hook script was executed.  This patch
breaks the reference cycle by using a WeakKeyDictionary instead of a
dict to track the objects.