PTL and TestFunctional fixes
These changes allow the suite of TestFunctional tests to run in a simple
single node environment using the following command.

    pbs_benchpress -F -t TestFunctional --follow-child

While some tests still fail, and many are skipped, the failures no
longer cascade.  Without these modifications, a test leaving the server,
a MoM or a comm in a bad state could result in the failure of all
subsequent tests.