Node state not updated at start of degraded reservation
If a node in a confirmed reservation goes down or offline, and the
reservation is not reconfirmed until the reservation starts, the state
of the nodes are not updated if resource assignments changed.  In one
particular case, nodes assigned exclusively to a starting reservation
that is degraded may be left in the free state rather than transitioned
into the resv-exclusive state.

This patch contains a fix and a test for the above scenario.  It also
includes a fix for a race condition identified by the existing
test_resv_server_restart test.